Remove Context interface
There is only a single usage by the Rule class so there is no value having
this extracted into an interface.
